Create FUNDING.yml
The EDSP ECO foundation is a non-profit organizations and works together with scientists and public & private organisations to develop innovative, durable and sustainable technologies and solutions for the transition towards a circular economy. EDSP ECO was founded by environmentalists with a vision of creating a simple way for anyone to contribute their eco-commitment. Our projects are funded bij EDSP (IT/NL), by donations and our patreons.